+++ src/shared/constants.ts
@@ -10,6 +10,7 @@ import { NpmPackage } from './npmPackage'
 export const extensionSettingsPrefix: string = 'aws'
 export const regionSettingKey: string = 'region'
 export const profileSettingKey: string = 'profile'
+export const mostRecentVersionKey: string = 'awsToolkitMostRecentVersion'
 
 export const hostedFilesBaseUrl: string = 'https://d3rrggjwfhwld2.cloudfront.net/'
 export const endpointsFileUrl: string = 'https://aws-toolkit-endpoints.s3.amazonaws.com/endpoints.json'
@@ -18,6 +19,7 @@ export const samAboutInstallUrl: string = 'https://aws.amazon.com/serverless/sam
 // tslint:disable-next-line:max-line-length
 export const vscodeMarketplaceUrl: string = 'https://marketplace.visualstudio.com/items?itemName=AmazonWebServices.aws-toolkit-vscode'
 export const githubUrl: string = 'https://github.com/aws/aws-toolkit-vscode'
+export const reportIssueUrl = `${githubUrl}/issues/new/choose`
 export const documentationUrl: string = 'https://docs.aws.amazon.com/toolkit-for-vscode/latest/userguide/welcome.html'
 // tslint:disable-next-line:max-line-length
 export const credentialHelpUrl: string = 'https://docs.aws.amazon.com/toolkit-for-vscode/latest/userguide/setup-credentials.html'