+++ src/shared/extensionUtilities.ts
@@ -6,8 +6,14 @@
 import * as _ from 'lodash'
 import * as path from 'path'
 import * as vscode from 'vscode'
+import * as nls from 'vscode-nls'
 import { ScriptResource } from '../lambda/models/scriptResource'
 import { ext } from '../shared/extensionGlobals'
+import { mostRecentVersionKey, pluginVersion } from './constants'
+import { readFileAsString } from './filesystemUtilities'
+import { Logger } from './logger'
+
+const localize = nls.loadMessageBundle()
 
 export class ExtensionUtilities {
     public static getLibrariesForHtml(names: string[]): ScriptResource[] {
@@ -68,4 +74,134 @@ export function safeGet<O, T>(obj: O | undefined, getFn: (x: O) => T): T | undef
     }
 
     return undefined
+}
+
+/**
+ * Helper function to show a webview containing the quick start page
+ *
+ * @param context VS Code Extension Context
+ */
+export async function showQuickStartWebview(context: vscode.ExtensionContext): Promise<void> {
+    try {
+        const view = await createQuickStartWebview(context)
+        view.reveal()
+    } catch {
+        vscode.window.showErrorMessage(
+            localize(
+                'AWS.command.quickStart.error',
+                'There was an error retrieving the Quick Start page'
+            )
+        )
+    }
+}
+
+/**
+ * Helper function to create a webview containing the quick start page
+ * Returns an unfocused vscode.WebviewPanel if the quick start page is renderable.
+ *
+ * @param context VS Code Extension Context
+ * @param page Page to load (use for testing); default: `quickStart.html`
+ */
+export async function createQuickStartWebview(
+    context: vscode.ExtensionContext,
+    page: string = 'quickStart.html'
+): Promise<vscode.WebviewPanel> {
+    const html = convertExtensionRootTokensToPath(
+        await readFileAsString(path.join(context.extensionPath, page)),
+        context.extensionPath
+    )
+    // create hidden webview, leave it up to the caller to show
+    const view = vscode.window.createWebviewPanel(
+        'html',
+        localize('AWS.command.quickStart.title', 'AWS Toolkit - Quick Start'),
+        { viewColumn: vscode.ViewColumn.Active, preserveFocus: true }
+    )
+    view.webview.html = html
+
+    return view
+}
+
+/**
+ * Utility function to search for tokens in a string and convert them to relative paths parseable by VS Code
+ * Useful for converting HTML images to webview-usable images
+ *
+ * @param text Text to scan
+ * @param basePath Extension path (from extension context)
+ */
+function convertExtensionRootTokensToPath(
+    text: string,
+    basePath: string
+): string {
+    return text.replace(/!!EXTENSIONROOT!!/g, `vscode-resource:${basePath}`)
+}
+
+/**
+ * Utility function to determine if the extension version has changed between activations
+ * False (versions are identical) if version key exists in global state and matches the current version
+ * True (versions are different) if any of the above aren't true
+ *
+ * TODO: Change the threshold on which we display the welcome page?
+ * For instance, if we start building nightlies, only show page for significant updates?
+ *
+ * @param context VS Code Extension Context
+ * @param currVersion Current version to compare stored most recent version against (useful for tests)
+ */
+export function isDifferentVersion(context: vscode.ExtensionContext, currVersion: string = pluginVersion): boolean {
+    const mostRecentVersion = context.globalState.get<string>(mostRecentVersionKey)
+    if (mostRecentVersion && mostRecentVersion === currVersion) {
+        return false
+    }
+
+    return true
+}
+
+/**
+ * Utility function to update the most recently used extension version
+ * Pulls from package.json
+ *
+ * @param context VS Code Extension Context
+ */
+export function setMostRecentVersion(context: vscode.ExtensionContext): void {
+    context.globalState.update(mostRecentVersionKey, pluginVersion)
+}
+
+/**
+ * Publishes a toast with a link to the welcome page
+ */
+async function promptQuickStart(): Promise<void> {
+    const view = localize(
+        'AWS.command.quickStart',
+        'View Quick Start'
+    )
+    const prompt = await vscode.window.showInformationMessage(
+        localize(
+            'AWS.message.prompt.quickStart.toastMessage',
+            'You are now using the AWS Toolkit for Visual Studio Code, version {0}',
+            pluginVersion
+        ),
+        view
+    )
+    if (prompt === view) {
+        vscode.commands.executeCommand('aws.quickStart')
+    }
+}
+
+/**
+ * Checks if a user is new to this version
+ * If so, pops a toast with a link to a quick start page
+ *
+ * @param context VS Code Extension Context
+ */
+export function toastNewUser(context: vscode.ExtensionContext, logger: Logger): void {
+    try {
+        if (isDifferentVersion(context)) {
+            setMostRecentVersion(context)
+            // the welcome toast should be nonblocking.
+            // tslint:disable-next-line: no-floating-promises
+            promptQuickStart()
+        }
+    } catch (err) {
+        // swallow error and don't block extension load
+        logger.error(err as Error)
+    }
 }