+++ src/test/templates/sam/samTemplateGenerator.test.ts
@@ -17,6 +17,8 @@ describe('SamTemplateGenerator', () => {
     const sampleCodeUriValue: string = 'sampleCodeUri'
     const sampleFunctionHandlerValue: string = 'sampleFunctionHandler'
     const sampleResourceNameValue: string = 'sampleResourceName'
+    const sampleMemorySize: number = 256
+    const sampleTimeout: number = 321
     const sampleRuntimeValue: string = 'sampleRuntime'
     const sampleEnvironment: CloudFormation.Environment = {
         Variables: {
@@ -35,13 +37,16 @@ describe('SamTemplateGenerator', () => {
         await del([tempFolder], { force: true })
     })
 
-    it('Produces a minimal template', async () => {
-        await new SamTemplateGenerator()
+    function makeMinimalTemplate(): SamTemplateGenerator {
+        return new SamTemplateGenerator()
             .withCodeUri(sampleCodeUriValue)
             .withFunctionHandler(sampleFunctionHandlerValue)
             .withRuntime(sampleRuntimeValue)
             .withResourceName(sampleResourceNameValue)
-            .withEnvironment(sampleEnvironment)
+    }
+
+    it('Produces a minimal template', async () => {
+        await makeMinimalTemplate()
             .generate(templateFilename)
 
         assert.strictEqual(await SystemUtilities.fileExists(templateFilename), true)
@@ -55,7 +60,79 @@ describe('SamTemplateGenerator', () => {
         assert.strictEqual(resource!.Properties!.CodeUri, sampleCodeUriValue)
         assert.strictEqual(resource!.Properties!.Handler, sampleFunctionHandlerValue)
         assert.strictEqual(resource!.Properties!.Runtime, sampleRuntimeValue)
+    })
+
+    it('Produces a template containing MemorySize', async () => {
+        await makeMinimalTemplate()
+            .withMemorySize(sampleMemorySize)
+            .generate(templateFilename)
+
+        assert.strictEqual(await SystemUtilities.fileExists(templateFilename), true)
+
+        const template: CloudFormation.Template = await CloudFormation.load(templateFilename)
+        assert.ok(template.Resources)
+        assert.notStrictEqual(Object.keys(template.Resources!).length, 0)
+
+        const resource = template.Resources![sampleResourceNameValue]
+        assert.ok(resource)
+        assert.strictEqual(resource!.Properties!.MemorySize, sampleMemorySize)
+    })
+
+    it('Produces a template containing Timeout', async () => {
+        await makeMinimalTemplate()
+            .withTimeout(sampleTimeout)
+            .generate(templateFilename)
+
+        assert.strictEqual(await SystemUtilities.fileExists(templateFilename), true)
+
+        const template: CloudFormation.Template = await CloudFormation.load(templateFilename)
+        assert.ok(template.Resources)
+        assert.notStrictEqual(Object.keys(template.Resources!).length, 0)
+
+        const resource = template.Resources![sampleResourceNameValue]
+        assert.ok(resource)
+        assert.strictEqual(resource!.Properties!.Timeout, sampleTimeout)
+    })
+
+    it('Produces a template containing Environment', async () => {
+        await makeMinimalTemplate()
+            .withEnvironment(sampleEnvironment)
+            .generate(templateFilename)
+
+        assert.strictEqual(await SystemUtilities.fileExists(templateFilename), true)
+
+        const template: CloudFormation.Template = await CloudFormation.load(templateFilename)
+        assert.ok(template.Resources)
+        assert.notStrictEqual(Object.keys(template.Resources!).length, 0)
+
+        const resource = template.Resources![sampleResourceNameValue]
+        assert.ok(resource)
         assert.deepStrictEqual(resource!.Properties!.Environment, sampleEnvironment)
+    })
+
+    it('Produces a template with a Globals section', async () => {
+        await makeMinimalTemplate()
+            .withGlobals({
+                Function: {
+                    Timeout: 5
+                }
+            })
+            .generate(templateFilename)
+
+        assert.strictEqual(await SystemUtilities.fileExists(templateFilename), true)
+
+        const template: CloudFormation.Template = await CloudFormation.load(templateFilename)
+        assert.ok(template.Globals, 'Expected loaded template to have a Globals section')
+        // tslint:disable:no-unsafe-any -- we don't care about the schema of globals for the test
+        const globals = template.Globals!
+        assert.notStrictEqual(Object.keys(globals).length, 0, 'Expected Template Globals to be not empty')
+
+        const functionKey = 'Function'
+        const timeoutKey = 'Timeout'
+        assert.ok(globals[functionKey], 'Expected Globals to contain Function')
+        assert.ok(globals[functionKey][timeoutKey], 'Expected Globals.Function to contain Timeout')
+        assert.strictEqual(globals[functionKey][timeoutKey], 5, 'Unexpected Globals.Function.Timeout value')
+        // tslint:enable:no-unsafe-any
     })
 
     it('errs if resource name is missing', async () => {